fix: Replaced returned type with SearchPathW (more bits)

+++ Windows-Native/System/Filesystem/Path.h
@@ -19,7 +19,7 @@ struct Path
     NTSTATUS(*RtlDosSearchPath_Ustr)(ULONG Flags,PUNICODE_STRING PathString,PUNICODE_STRING FileNameString,PUNICODE_STRING ExtensionString,PUNICODE_STRING CallerBuffer,PUNICODE_STRING DynamicString,PUNICODE_STRING* FullNameOut,PSIZE_T FilePartSize,PSIZE_T LengthNeeded);
     NTSTATUS(*RtlDosPathNameToNtPathName_U)(PCWSTR DosFileName, PUNICODE_STRING NtFileName, PWSTR* FilePart, PRTL_RELATIVE_NAME_U RelativeName);
     NTSTATUS(*RtlGetExePath)(PCWSTR name, PWSTR* path);
-    DWORD(*SearchPathW)(LPCWSTR lpPath, LPCWSTR lpFileName, LPCWSTR lpExtension, DWORD nBufferLength, LPWSTR lpBuffer, LPWSTR* lpFilePart);
+    SIZE_T(*SearchPathW)(LPCWSTR lpPath, LPCWSTR lpFileName, LPCWSTR lpExtension, DWORD nBufferLength, LPWSTR lpBuffer, LPWSTR* lpFilePart);
     NTSTATUS(*RtlGetSearchPath)(PWCHAR* SearchPath);
 };
